Add serial printing support
This allows us to communicate with the outside world! We can log and
print errors and everything---as long as we don't cause a CPU exception
of course, because then we still triple-fault. This happened when
writing this code, because the ELF was mapped read-only. Now everything
is wrapped read-write, but this is perhaps also not ideal.