Allow virtual/physical size mismatch in ELF
This will soon be necessary because of the static variables we want to
use for our logger. While we're at it, we use information from the ELF
to select the desired page table flags.

When the virtual length is larger than the physical length, the relevant
information from the last frame is copied to a fresh frame, which is
mapped to the desired page. If necessary, additional fresh frames are
allocated and mapped to subsequent pages. All memory in the newly
allocated frames beyond the copied data is zeroed out.