Handle CPU exceptions
Handle the following:
- Breakpoints (int3) for easy debugging
- Page faults
- Double faults (with seperate stack)

Some extra testing is still needed, but even now it's not working
perfectly. A panic is introduced in the breakpoint handler because
without it a general protection fault (resulting in a double fault
because general protection faults aren't handled at the moment) happens
and I don't know why.