Reset stack-segment register to fix iretq
An iretq from the breakpoint handler used to result in a general
protection fault with error code 0x30. This appears to be because the
stack-segment register has leftover data from the UEFI firmware, as it
does not happen when BIOS boot is used using the bootloader crate. The
following is an overview of the contents of all segment registers with
different booting methods at the iretq of the breakpoint handler:

| Segment    | With       | With |
| Registers  | bootloader | UEFI |
+------------+------------+------+
| cs         | 0x8        | 0x8  |
| ss, fs, gs | 0x0        | 0x30 |
| ds, es     | 0x10       | 0x30 |

Except for the code-segment selector (which is explicitly set when the
new GDT is loaded) they are different, presumably because of the booting
method. For now, the solution is to simply put 0 in the stack-segment
register and that prevents the general protection fault, but a more
correct solution may need to be found in the future.