Add PIC8259 support and timer interrupts
This is basically just following the tutorial by Philipp Oppermann. The
most major change is that UEFI seems to mask all PIC8259 interrupts by
default, so they need to be unmasked before they can be used. The API of
the crate does not provide a way for us to do this, so we directly write
to the relevant ports (handling this through a nice abstraction is a
problem for the future). The masks are the ones present when a BIOS boot
is performed; it's possible these masks need to be changed if it is
desired to look at different interrupts.

Timer interrupts are aplenty, so only every 1000th occurrence is logged.
The first one happens right around printing the header with its empty
lines of the operating system, so this is now done in a single println
call to prevent the timer log from also being between the blank lines.