Add rudimentary userspace
Just the absolute basics. Additional GDT entries; an actual segment
register, and syscall/sysret can be enabled. Then, we can switch to
userspace with sysretq and just loop there...

Interrupts are disabled, as enabling interrupts causes a double
fault (because first a page fault is raised for some reason) after a
timer interrupt. All mappings done by the UEFI stub are made user
accessible for an easy start, but ideally there'd be a separate area in
virtual memory for userspace such that only those pages need to be user
accessible.