Make kernel relocatable
For now, just support R_X86_64_RELATIVE entries in RELA64 sections.
Other types yield errors and/or warnings and can be added on an
as-needed basis. The "-pie" link argument is now passed through the
build script as it is the same for all invocations of the same
target (otherwise, because it is done through RUSTFLAGS, core and
dependencies would be rebuilt each time).