Make sysret/syscall non-terminal
Instead, jump back to where sysret was invoked. Doing this in only some
cases, and otherwise jumping back to userspace is next and allows for
useful syscalls that aren't exit.

This code uses a label in inline assembly to find the point to jump back
to. While this works it may yield errors later on because the compiler
is allowed to duplicate assembly blocks which would duplicate the label
as well. Also, a bunch of lateout flags are required to mark all
registers as clobbered.