Make non-terminating syscalls
By looping while handling syscalls and saving rsp/rip it's possible to
perform syscalls that don't just end the program. Now we can let the
kernel print a message! Soon the dummy userspace crate may become a
misnomer...

Note that in terms of safety the kernel isn't great yet: it's quite
trusting in userspace, both that it will call syscall 0 at some point to
pass control back to the kernel and that it does not pass malicious
parameters to the printing function (e.g. printing kernel memory or
creating a too-long slice or causing a page fault).