Allow release optimizations
The --release flag has to be done before the subcommand :/, but
restructuring that nicely is a problem for the future.

Currently, this breaks the kernel because the memory map gets
corrupted. To be fair, the memory map turns out to be corrupted even if
--release isn't used but then the corruption happens in a way that
doesn't break the kernel. Who knew that trying to unsafely get a slice
that doesn't exist as such would give such problems? Anyway, this may
mean that we need to copy the memory descriptors to structures of our
own or put the UEFI crate's iterator in the boot information directly.